+++ IN505/mesformes/Triangle.h
@@ -9,11 +9,12 @@ class Triangle:public Point
 		Point p1;
 		Point p2;
 		Point p3;	
+		
 	public:
 		Triangle();
 		Triangle(Point p1, Point p2, Point p3);
 		Triangle(Triangle& pc);
-		void deplacer(double dx, double dy);
+		void deplacer(int dx, int dy);
 		void afficher();
 		void cloner(const Point& p);
 		~Triangle();
